requirements.txt
certifi
charset-normalizer
idna
pyTelegramBotAPI
requests
urllib3
google-auth-oauthlib
google-auth
google-api-python-client
httplib2
